+++ tests/docker/runners/bash/all-features.sh
@@ -12,7 +12,7 @@ done
 
 FEATURES="network-logs pool-prefer-active custom-reconnect-errors ignore-auth-error serde-json blocking-encoding
           full-tracing reconnect-on-auth-error monitor metrics sentinel-client subscriber-client dns debug-ids
-          check-unresponsive replicas client-tracking"
+          check-unresponsive replicas client-tracking codec"
 
 if [ -z "$FRED_CI_NEXTEST" ]; then
   cargo test --release --lib --tests --features "$FEATURES" -- --test-threads=1 "$@"